Initial approach for Rails 3 compatibility
