+++ app/models/translation.rb
