+++ app/models/translation_option.rb
