+++ app/views/locales/edit.html.erb
