+++ app/views/locales/index.html.erb
