+++ app/views/locales/new.html.erb
