+++ app/views/locales/show.html.erb
