+++ app/views/translations/asset_translations.html.erb
