+++ app/views/translations/index.html.erb
