+++ app/views/translations/show.html.erb
