+++ init.rb
@@ -1,1 +1,2 @@
 require 'i18n_backend_database'
+require 'i18n_backend_database/backend_config'
