+++ lib/i18n_backend_database/backend_config.rb
@@ -1,0 +1,8 @@
+# This is where the i18n backend handler is overriden
+module I18n
+  Config.class_eval do
+    def backend
+      @@backend ||= Backend::Database.new
+    end
+  end
+end
